Update to v1.13.10 stable

+++ src/main/java/com/github/manolo8/darkbot/gui/utils/TableDoubleEditor.java
@@ -30,7 +30,7 @@ public class TableDoubleEditor extends DefaultCellEditor {
 
     public boolean stopCellEditing() {
         try {
-            value = nf.parse((String) super.getCellEditorValue());
+            value = nf.parse((String) super.getCellEditorValue()).doubleValue();
         } catch (Exception e) {
             ((JComponent)getComponent()).setBorder(new LineBorder(Color.red));
             return false;
